requirements.txt
python-telegram-bot==20.3
APScheduler==3.11.0
Flask==2.3.3
httpx==0.24.1
httpcore==0.17.3
certifi==2025.4.26
h11==0.14.0
idna==3.10
sniffio==1.3.1
tzlocal==5.3.1
typing_extensions==4.13.2
gunicorn==21.2.0